requirements.txt
# Core
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0

# LLM
google-genai>=1.0
openai>=1.0                  # Grok uses OpenAI-compatible API

# Review Scraping
google-play-scraper>=1.2

# MCP
mcp>=1.0

# PDF Generation
reportlab>=4.0

# Backend API
fastapi>=0.110
uvicorn>=0.27

# Database
# (sqlite3 is part of Python stdlib)

# Testing
pytest>=8.0
pytest-asyncio>=0.23
langdetect>=1.0.9
